noch nicht ganz ausbalanciert

+++ include/noiseutils.h
@@ -1209,7 +1209,7 @@ namespace noise
         /// Constructor.
         WriterTER ():
           m_pSourceNoiseMap (NULL),
-          m_metersPerPoint (DEFAULT_METERS_PER_POINT)
+          m_metersPerPoint ((const float)DEFAULT_METERS_PER_POINT)
         {
         }
